Split out objs into single files. Clean up. TODOs
