+++ packages/docs-markdown/package.json
@@ -11,7 +11,7 @@
 		"url": "https://github.com/Microsoft/vscode-docs-authoring/issues"
 	},
 	"engines": {
-		"vscode": "^1.53.0"
+		"vscode": "^1.55.0"
 	},
 	"repository": {
 		"url": "https://github.com/Microsoft/vscode-docs-authoring.git"
@@ -800,7 +800,7 @@
 		"@types/mocha": "^7.0.2",
 		"@types/node": "^12.20.6",
 		"@types/recursive-readdir": "^2.2.0",
-		"@types/vscode": "^1.53.0",
+		"@types/vscode": "^1.55.0",
 		"@types/yamljs": "^0.2.30",
 		"@typescript-eslint/eslint-plugin": "^2.34.0",
 		"@typescript-eslint/eslint-plugin-tslint": "^2.34.0",
