+++ packages/docs-markdown/src/controllers/doc-index/dependent-operation-enum.ts
@@ -1,0 +1,6 @@
+export enum DependentOperationEnum {
+	DirectChildren,
+	AllChildBlocks,
+	CurrentBlocks,
+	AllBlocks
+}
