+++ packages/docs-markdown/src/controllers/doc-index/doc-index-controller.ts
@@ -8,7 +8,8 @@ import {
 } from '../../helper/common';
 import { format } from '../../helper/format';
 import { sendTelemetryData } from '../../helper/telemetry';
-import { AuditRule, AuditEntry } from './audit';
+import { AuditEntry } from './audit-entry';
+import { AuditRule } from './audit-rule';
 import * as vscode from 'vscode';
 import { ContentMatch } from './content-match';
 import { ContentBlock } from './content-block';
@@ -18,11 +19,11 @@ let commandOption: string;
 
 export function insertDocIndexCommand() {
 	return [
-        {
-            command: verify.name,
-            callback: verify
-        }
-    ];
+		{
+			command: verify.name,
+			callback: verify
+		}
+	];
 }
 
 /**
@@ -30,8 +31,8 @@ export function insertDocIndexCommand() {
  */
 export function verify() {
 	const editor = vscode.window.activeTextEditor;
-	AuditRule.LoadRules(JSON.parse(readFileSync('./AuditRules.json', 'utf-8')));
-	
+	AuditRule.LoadRules(<AuditRule[]>JSON.parse(readFileSync('./audit-rules.json', 'utf-8')));
+
 	if (!editor) {
 		noActiveEditorMessage();
 		return;
@@ -47,28 +48,27 @@ export function verify() {
 		let metadata = ContentMatch.readMetadata(metadataString);
 		if (metadata.keys.length == 0) {
 			metadata = ContentMatch.extractMetadata(entireFile);
-		}		
-		let topic = "";
-		if (metadata.has("ms.topic")) {
-			topic = metadata["ms.topic"];
 		}
-
-		if (topic == "") {
-			vscode.window.activeTerminal.sendText("No MS.Topic detected for MVC guidance")
-		}
-		else {
-			let rules = AuditRule.Rules.filter(e => e.ruleGroup == "MVC" && e.ruleSet.toLowerCase() == topic.toLowerCase() && e.dependsOn == undefined); 
+		let topic = metadata.has('ms.topic') ? metadata['ms.topic'] : '';
+		if (topic == '') {
+			vscode.window.activeTerminal.sendText('No MS.Topic detected for MVC guidance');
+		} else {
+			let rules = AuditRule.Rules.filter(
+				e =>
+					e.ruleGroup == 'MVC' &&
+					e.ruleSet.toLowerCase() == topic.toLowerCase() &&
+					e.dependsOn == undefined
+			);
 			if (rules.length == 0) {
-				vscode.window.activeTerminal.sendText("No MVC Guidance for " + topic);
-			}
-			else {				
-				for (let i = 0; i < rules.length;i++) {
+				vscode.window.activeTerminal.sendText('No MVC Guidance for ' + topic);
+			} else {
+				for (let i = 0; i < rules.length; i++) {
 					let theseAudits = rules[i].test(blocks, fileName, metadata, entireFile, blocks);
 					for (let j = 0; j < theseAudits.length; j++) {
 						vscode.window.activeTerminal.sendText(theseAudits[j].title);
 					}
 				}
 			}
-		}		
+		}
 	}
 }
