+++ packages/docs-markdown/src/controllers/doc-index/logical-operator-enum.ts
@@ -1,0 +1,5 @@
+export enum LogicalOperatorEnum {
+	Or,
+	And,
+	None
+}
