+++ packages/docs-markdown/src/controllers/doc-index/on-failure-enum.ts
@@ -1,0 +1,4 @@
+export enum OnFailureEnum {
+	SkipDependents,
+	SummarizeDependents
+}
