+++ packages/docs-markdown/src/controllers/doc-index/rule-set-type-enum.ts
@@ -1,0 +1,5 @@
+export enum RuleSetTypeEnum {
+	Horizontals,
+	MVC,
+	None
+}
