+++ packages/docs-markdown/src/controllers/doc-index/siblings-in-enum.ts
@@ -1,0 +1,8 @@
+export enum SiblingsInEnum {
+	SameParent,
+	SameHeader,
+	SameFile,
+	SameNode,
+	SameTopNode,
+	None
+}
