+++ packages/docs-markdown/src/controllers/doc-index/store-result-enum.ts
@@ -1,0 +1,6 @@
+export enum StoreResultEnum {
+	OnSuccess,
+	OnSuccessOrFailure,
+	OnFailure,
+	DoNotStore
+}
